requirements.txt
apify-client>=1.6.0
supabase>=2.3.4
openai>=1.14.0
python-dotenv>=1.0.1
modal>=0.60.0
